chore: move api extractor
