--- api-extractor/api-extractor-ui.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://developer.microsoft.com/json-schemas/api-extractor/v7/api-extractor.schema.json",
-  "extends": "../api-extractor.json",
-
-  "mainEntryPointFilePath": "<projectFolder>/dist/esm/ui/index.d.ts",
-
-  "apiReport": {
-    "enabled": true,
-    "reportFolder": "<projectFolder>/api-extractor/reports/",
-    "reportFileName": "ui"
-  }
-}
