+++ package.json
@@ -67,20 +67,20 @@
     "inspect": "pnpm run build && mcp-inspector -- dist/esm/index.js",
     "prettier": "prettier",
     "check": "concurrently \"pnpm run build\" \"pnpm run check:types\" \"pnpm run check:lint\" \"pnpm run check:format\" \"pnpm run check:dependencies\" && pnpm run check:api",
-    "check:api-report": "concurrently \"api-extractor run --config api-extractor.json\" \"api-extractor run --config api-extractor/api-extractor-web.json\" \"api-extractor run --config api-extractor/api-extractor-tools.json\"",
+    "check:api-report": "concurrently \"api-extractor run --config api-extractor.json\" \"api-extractor run --config api-extractor/api-extractor-web.json\" \"api-extractor run --config api-extractor/api-extractor-tools.json\" \"pnpm --filter @mongodb-js/mcp-ui check:api\"",
     "check:api": "pnpm run check:api-report || (printf \"\\nPublic API check failed. If you intended to change the public API, run:\\n\\n  pnpm run update:api\\n\" && exit 1)",
     "check:lint": "eslint .",
     "check:dependencies": "knip --strict",
     "check:format": "prettier -c .",
-    "update:api": "pnpm run build && concurrently \"api-extractor run --local --config api-extractor.json\" \"api-extractor run --local --config api-extractor/api-extractor-web.json\" \"api-extractor run --local --config api-extractor/api-extractor-tools.json\"",
+    "update:api": "pnpm run build && concurrently \"api-extractor run --local --config api-extractor.json\" \"api-extractor run --local --config api-extractor/api-extractor-web.json\" \"api-extractor run --local --config api-extractor/api-extractor-tools.json\" \"pnpm --filter @mongodb-js/mcp-ui update:api\"",
     "check:types": "tsc --noEmit --project tsconfig.json",
     "fix": "pnpm run fix:lint && pnpm run reformat",
     "fix:lint": "eslint . --fix",
     "reformat": "prettier --write .",
     "generate": "pnpm run generate:api && pnpm run generate:arguments && pnpm run generate:ui",
-    "generate:ui": "pnpm --filter @mongodb-js/mcp-ui generate:ui",
+    "generate:ui": "pnpm --filter @mongodb-js/mcp-scripts generate:ui",
     "generate:api": "./scripts/generate.sh",
-    "generate:arguments": "tsx --tsconfig scripts/generate/tsconfig.json scripts/generate/index.ts",
+    "generate:arguments": "pnpm --filter @mongodb-js/mcp-scripts generate",
     "pretest": "pnpm run build",
     "test": "vitest --project eslint-rules --project unit-and-integration --coverage --run",
     "test:accuracy": "sh ./scripts/accuracy/runAccuracyTests.sh",
