+++ packages/scripts/README.md
