+++ packages/scripts/src/generateUI.ts
@@ -11,7 +11,7 @@ import { fileURLToPath } from "url";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-const rootDir = join(__dirname, "..");
+const rootDir = join(__dirname, "..", "..", "ui");
 
 export function generateUI(): void {
     console.log("Building UI modules...");
