--- packages/ui/scripts/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../../tsconfig.build.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "noEmit": true
-  },
-  "include": ["**/*.ts"]
-}
