--- scripts/generate/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "rootDir": ".",
-    "noEmit": true,
-    "paths": {
-      "ts-levenshtein": ["../node_modules/ts-levenshtein/dist/index.d.mts"]
-    }
-  },
-  "include": ["**/*.ts"]
-}
